Change sqrt to sqrtf

+++ cuda/2PCFisotropic/normal/kernels.cuh
@@ -38,7 +38,7 @@ __device__ void count_distances11(double *XX, PointW3D *elements, int start, int
             w2 = elements[j].w;
             d = (x2-x1)*(x2-x1)+(y2-y1)*(y2-y1)+(z2-z1)*(z2-z1);
             if (d<=dd_max){
-                bin = (int)(sqrt(d)*ds);
+                bin = (int)(sqrtf(d)*ds);
                 v = sum*w1*w2;
                 atomicAdd(&XX[bin],v);
             }
@@ -80,7 +80,7 @@ __device__ void count_distances12(double *XX, PointW3D *elements, int start1, in
             w2 = elements[j].w;
             d = (x2-x1)*(x2-x1)+(y2-y1)*(y2-y1)+(z2-z1)*(z2-z1);
             if (d<=dd_max){
-                bin = (int)(sqrt(d)*ds);
+                bin = (int)(sqrtf(d)*ds);
                 v = sum*w1*w2;
                 atomicAdd(&XX[bin],v);
             }
@@ -123,7 +123,7 @@ __device__ void count_distancesXY(double *XY, PointW3D *elements1, int start1, i
             w2 = elements2[j].w;
             d = (x2-x1)*(x2-x1)+(y2-y1)*(y2-y1)+(z2-z1)*(z2-z1);
             if (d<=dd_max){
-                bin = (int)(sqrt(d)*ds);
+                bin = (int)(sqrtf(d)*ds);
                 v = sum*w1*w2;
                 atomicAdd(&XY[bin],v);
             }
@@ -159,7 +159,7 @@ __global__ void make_histoXX(double *XX, PointW3D *elements, DNode *nodeD, int p
 
             float ds = ((float)(bn))/dmax, dd_max=dmax*dmax;
             float nx1=nodeD[idx].nodepos.x, ny1=nodeD[idx].nodepos.y, nz1=nodeD[idx].nodepos.z;
-            float d_max_node = dmax + size_node*sqrt(3.0);
+            float d_max_node = dmax + size_node*sqrtf(3.0);
             d_max_node*=d_max_node;
 
             // Counts distances within the same node
@@ -241,7 +241,7 @@ __global__ void make_histoXY(double *XY, PointW3D *elementsD, DNode *nodeD, Poin
 
             float ds = ((float)(bn))/dmax, dd_max=dmax*dmax;
             float nx1=nodeD[idx].nodepos.x, ny1=nodeD[idx].nodepos.y, nz1=nodeD[idx].nodepos.z;
-            float d_max_node = dmax + size_node*sqrt(3.0);
+            float d_max_node = dmax + size_node*sqrtf(3.0);
             d_max_node*=d_max_node;
             
             int idx2,u,v,w; //Position of the second node
